Move code that updates document db metrics into separate class.
