+++ searchcore/src/vespa/searchcore/proton/server/CMakeLists.txt
@@ -28,6 +28,7 @@ vespa_add_library(searchcore_server STATIC
     documentbucketmover.cpp
     documentdb.cpp
     documentdb_commit_job.cpp
+    documentdb_metrics_updater.cpp
     documentdbconfig.cpp
     documentdbconfigscout.cpp
     documentdbconfigmanager.cpp
