+++ searchcore/src/vespa/searchcore/proton/server/document_db_explorer.cpp
@@ -32,10 +32,7 @@ DocumentDBExplorer::get_state(const Inserter &inserter, bool full) const
         StateReporterUtils::convertToSlime(*_docDb->reportStatus(), ObjectInserter(object, "status"));
     }
     {
-        // TODO(geirst): Avoid const cast by adding const interface to
-        // IDocumentMetaStoreContext as seen from IDocumentSubDB.
-        DocumentMetaStoreReadGuards dmss
-            (const_cast<DocumentSubDBCollection &>(_docDb->getDocumentSubDBs()));
+        DocumentMetaStoreReadGuards dmss(_docDb->getDocumentSubDBs());
         Cursor &documents = object.setObject("documents");
         documents.setLong("active", dmss.numActiveDocs());
         documents.setLong("ready", dmss.numReadyDocs());
