+++ searchcore/src/vespa/searchcore/proton/test/dummy_document_sub_db.h
@@ -67,6 +67,7 @@ struct DummyDocumentSubDb : public IDocumentSubDB
     const ISummaryAdapter::SP &getSummaryAdapter() const override { return _summaryAdapter; }
     const IIndexWriter::SP &getIndexWriter() const override { return _indexWriter; }
     IDocumentMetaStoreContext &getDocumentMetaStoreContext() override { return _metaStoreCtx; }
+    const IDocumentMetaStoreContext &getDocumentMetaStoreContext() const override { return _metaStoreCtx; }
     IFlushTargetList getFlushTargets() override { return IFlushTargetList(); }
     size_t getNumDocs() const override { return 0; }
     size_t getNumActiveDocs() const override { return 0; }
